changes to support setting OSC CTRL bits based on PCI config space info struct (#34)
## Description

The OSC method in the context of PCI conveys information to the
Operating System on which features it may take control of (native
control) and which features will be controlled by firmware
(firmware-first).

The previous version of this library uses a hard-coded OSC method that
does not allow the consumer to specify if certain features should be
granted to the OS or not.

This change uses the aml fixup libraries to set the CTRL bits in the OSC
method.

For each item, place an "x" in between `[` and `]` if true. Example:
`[x]`.
_(you can also check items in the GitHub UI)_

- [X] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested
Tested on platform that utilized DyamicTablesPkg. Verified that selected
osc template will be used in the PCI ACPI device when specified in the
PCI Config Space Info structure
## Integration Instructions
Consumers who utlize the DynamicTablesPkg should specify in
CM_ARM_PCI_CONFIG_SPACE_INFO.OscControlBitmap which OSC CTRL bits should
be set if the OS requests them

Co-authored-by: joelopez <joelopez@microsoft.com>