ArmPkg: Fix pointer type errors
The code tries to use a data pointer type for a function pointer and
Visual Studio doesn't like it.

Fixes this warning:
  nonstandard extension, function/data pointer conversion in expression

This change represents the function pointer with the appropriate type
(instead of VOID*) and defines the type with a typedef for code
clarity.

A minor change is added to also validate function pointer parameters
in an already modified function to generally improve its robustness.

Some practices like adding a function description are not added
because the function prototype is from ArmMmuLib which doesn't
currently have function descriptions and adding them is outside the
scope of the compiler fix being added in this change.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>