Revert "ArmPkg: GICv3: Fix ARM_GICD_IROUTER incorrect writes for SPIs"
This reverts commit f735cd4481511bc5dcfc9ac4afe10fb97e3528e4.

This commit references an older GICv3 spec (that cannot be found)
that had a different ARM_GICD_ROUTER definition at 0x6000 instead
of 0x6100. The current GICv3 spec defines ARM_GICD_ROUTER as
0x6100 [1].

It also was determined that the logic between this PR and the
existing edk2 logic was the same but Mu kept this behavior as
it was believed to better align to the spec, but again the
current spec says 0x6100, so this logic is dropped.

No known Mu consumer was using this definition, but it is marked
as a breaking change because if any consumer was using the
definition, it is now different.

Bugzilla ref: https://bugzilla.tianocore.org/show_bug.cgi?id=1294

[1] https://developer.arm.com/documentation/ihi0069/latest