ArmPlatformPkg: Carve out the mm buffer region from the bulk system memory hob (#431)
The MM communication buffer memory is being added in MM communication
DXE driver through gDS calls. This means that the region should not be
reported as available through resource descriptor hob.

This change will carve out the MM communication buffer from the system
memory.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

This was tested on QEMU SBSA and booted to OS desktop.

N/A